+++ source/ftrack_api/session.py
@@ -16,6 +16,7 @@ import hashlib
 import tempfile
 import threading
 import atexit
+import warnings
 
 import requests
 import requests.auth
@@ -1690,12 +1691,12 @@ class Session(object):
             if "entity_data" in data:
                 for key, value in data["entity_data"].items():
                     if isinstance(value, ftrack_api.entity.base.Entity):
-                        data["entity_data"][key] = self._entity_reference(value)
+                        data["entity_data"][key] = self.entity_reference(value)
 
             return data
 
         if isinstance(item, ftrack_api.entity.base.Entity):
-            data = self._entity_reference(item)
+            data = self.entity_reference(item)
 
             with self.auto_populating(True):
 
@@ -1723,7 +1724,7 @@ class Session(object):
                             attribute, ftrack_api.attribute.ReferenceAttribute
                         ):
                             if isinstance(value, ftrack_api.entity.base.Entity):
-                                value = self._entity_reference(value)
+                                value = self.entity_reference(value)
 
                         data[attribute.name] = value
 
@@ -1738,14 +1739,14 @@ class Session(object):
         if isinstance(item, ftrack_api.collection.Collection):
             data = []
             for entity in item:
-                data.append(self._entity_reference(entity))
+                data.append(self.entity_reference(entity))
 
             return data
 
         raise TypeError('{0!r} is not JSON serializable'.format(item))
 
-    def _entity_reference(self, entity):
-        '''Return reference to *entity*.
+    def entity_reference(self, entity):
+        '''Return entity reference that uniquely identifies *entity*.
 
         Return a mapping containing the __entity_type__ of the entity along with
         the key, value pairs that make up it's primary key.
@@ -1758,6 +1759,29 @@ class Session(object):
             reference.update(ftrack_api.inspection.primary_key(entity))
 
         return reference
+
+    def _entity_reference(self, entity):
+        '''Return entity reference that uniquely identifies *entity*.
+
+        Return a mapping containing the __entity_type__ of the entity along with
+        the key, value pairs that make up it's primary key.
+
+        .. note::
+
+            This private method is now available as public method
+            :meth:`entity_reference`. This alias remains for backwards
+            compatibility, but will be removed in version 2.0.
+
+        '''
+        warnings.warn(
+            (
+                "Session._entity_reference is now available as public method "
+                "Session.entity_reference. The private method will be removed "
+                "in version 2.0."
+            ),
+            PendingDeprecationWarning
+        )
+        return self.entity_reference(entity)
 
     def decode(self, string):
         '''Return decoded JSON *string* as Python object.'''